Make use of existing components and link to /member page. Fix query for sponsorships
