+++ frontend/src/styles/common.scss
@@ -160,13 +160,26 @@
         margin-top: 20px;
     }
 
-    img {
-        display: block;
-        width: 33%;
+    .image {
+        width: 200px; /* Set a fixed width */
+        height: 200px; /* Set a fixed height */
+        border-radius: 50%; /* Make it circular */
+        overflow: hidden; /* Hide anything outside the circle */
+        display: flex;
+        justify-content: center;
+        align-items: center;
         margin: 0 auto;
     }
 
-    p { margin: 20px }
+    img {
+        width: 100%; /* Makes the image cover the container width */
+        height: 100%; /* Makes the image cover the container height */
+        object-fit: cover; /* Ensures the image covers the area without stretching */
+    }
+
+    p {
+        margin: 20px;
+    }
 
     ul {
         margin-top: 0px;
